Evas textblock: add support for hyphenation modes
Hyphenation is the way for the text to do line breaking by breaking
that word at the end of the line that doesn't fit, and appending a
hyphen ("‐" HYPHEN U+2010) to it, continuing the rest of it in the
following line.
This adds support for several modes of hyphenation in Textblock.

Most of the implementation changes are in the line wrapping code by
squeezing-in a hyphen text item at the end of the broken line, if the
wrapping was done due to hyphenation of a word.

The biggest part here was handling of AUTO mode, which uses external
dictionaries with Hunspell's hyphenation API. This was integrated in
the word and mixed wrap mode to search for hyphenation possibilities for a
better utilization of the line's width.

To enable support for auto-hyphenation, use the configure options:
  --enable-hyphen (requires "hyphen" library installed)
  --with-dictionaries-hyphen-dir=DIR (e.g. /usr/share/hyphen is wher
          it is installed in Arch Linux's "hyphen" package)

More info is provided in the eo doc.

@feature