Add Master/Slave type to sessions
This feature is needed for security issues. We need to be sure no
application can connect to another application.

Regular applications connect to a certain local socket, when debug tools
connect to another. In systems where the sockets can be used only by a
specific group of applications, it permits to separate applications and
debug tools.

The next steps are:
- During forwarding, check that packets are transmitted between a master
and a slave only (no slave/slave communication)
- Only master should be able to observe connections/disconnections.