Ignore SIGPIPE
Without it, a crash can happen on disconnection.

For example, if clouseau_client is run and connects to the local daemon,
two sessions are created: the slave one and the master one.

When clouseau terminates, the daemon receives a disconnection on the
slave socket, closes it, removes the session and informs the observers
(including the master connection of Clouseau) that this client is no more
present. Because the Clouseau is already disconnected, the daemon tries
to write to a socket that is already broken, leading to a crash.