eo: make callback,add and callback,del a hot event
This is absolutely necessary as there is no way to track callback other than
with this method and we would be blind to any callback registered during a freeze
(specifically when think interact in all out stack).

Also this should definitively be a patch on just the .eo file, but I can't
figure out how to do so.