Entry: Move theming size evaluation to eo_base_finalize
Entry inherits from layout, and we have sizing_eval() calls both entry
and the layout parent object, due to super calls in smart_add.
This is not only redundant but actually wrong.

On a side node, Super calls should be handled with care, and maybe in
the future we should remove them all from the smart_add functions and
do it properly.

Also, Label is going to inherit from Entry, so we need to skip the
theming of the object (and especially not do it in the super'd
smart_add).

To be continued with commits the follows for merging Label to Entry.