Evas textblock: add object_item API
The object item replaces the format item. Normal usage of adding items that
occupy space in the text is now blocked from annotation. This means that formats
like "<item ...>" are not supported via annotation_insert/set.
Instead, use object_item_insert to add items. Another difference is that
you specify at which position you will add the item (and not a range, like in
annotation API).

Essentially, this adds the OBJECT_REPLACEMENT_CHARACTER (OBJ) and adds a format
to it, like it normally did. Also, you can query its annotation and get the
attributes of the item like "size=32x32", "vsize=..." etc. Deleting it will also
delete the inserted OBJ character, and vice-versa. So this API is pretty
self-contained and shouldn't feel uncomfortable to use.

@feature